requirements.txt
wandb==0.18.7
datasets==3.1.0
nltk==3.9.1
numpy==2.1.3
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
rouge_score==0.1.2
tiktoken==0.8.0
torch==2.5.1+cu118
tqdm==4.66.5
